requirements.txt
fastapi
uvicorn
pillow
numpy
streamlit
